requirements.txt
fastapi
fastapi-utils
uvicorn[standard]
requests
PyJWT[crypto]
pymongo
openAI
python-dotenv